+++ app/Livewire/Erp/Production/Jobcard/Upsert.php
@@ -476,45 +476,48 @@ class Upsert extends Component
 
     public function save(): string
     {
-        if ($this->order_id != '') {
+        if (session()->has('tenant_id')) {
 
-            if ($this->vid == "") {
+            if ($this->order_id != '') {
 
-                $obj = Jobcard::create([
-                    'vno' => $this->vno,
-                    'vdate' => $this->vdate,
-                    'order_id' => $this->order_id,
-                    'style_id' => $this->style_id,
-                    'total_qty' => $this->total_qty,
-                    'active_id' => '1',
-                    'user_id' => \Auth::id(),
-                ]);
-                $this->saveItem($obj->id);
+                if ($this->vid == "") {
 
-                $message = "Saved";
+                    $obj = Jobcard::create([
+                        'vno' => $this->vno,
+                        'vdate' => $this->vdate,
+                        'order_id' => $this->order_id,
+                        'style_id' => $this->style_id,
+                        'total_qty' => $this->total_qty,
+                        'active_id' => '1',
+                        'user_id' => \Auth::id(),
+                    ]);
+                    $this->saveItem($obj->id);
 
-            } else {
-                $obj = Jobcard::find($this->vid);
-                $obj->vno = $this->vno;
-                $obj->vdate = $this->vdate;
-                $obj->order_id = $this->order_id;
-                $obj->style_id = $this->style_id;
-                $obj->total_qty = $this->total_qty;
-                $obj->active_id = '1';
-                $obj->user_id = \Auth::id();
-                $obj->save();
+                    $message = "Saved";
 
-                DB::table('jobcard_items')->where('jobcard_id', '=', $obj->id)->delete();
-                $this->saveItem($obj->id);
-                $message = "Updated";
+                } else {
+                    $obj = Jobcard::find($this->vid);
+                    $obj->vno = $this->vno;
+                    $obj->vdate = $this->vdate;
+                    $obj->order_id = $this->order_id;
+                    $obj->style_id = $this->style_id;
+                    $obj->total_qty = $this->total_qty;
+                    $obj->active_id = '1';
+                    $obj->user_id = \Auth::id();
+                    $obj->save();
+
+                    DB::table('jobcard_items')->where('jobcard_id', '=', $obj->id)->delete();
+                    $this->saveItem($obj->id);
+                    $message = "Updated";
+                }
+                $this->getRoute();
+                $this->vno = '';
+                $this->vdate = '';
+                $this->order_id = '';
+                $this->style_id = '';
+                $this->total_qty = '';
+                return $message;
             }
-            $this->getRoute();
-            $this->vno = '';
-            $this->vdate = '';
-            $this->order_id = '';
-            $this->style_id = '';
-            $this->total_qty = '';
-            return $message;
         }
         return '';
     }
